[alpaka] Split shared objects to backend-specific libraries
Move common code shared between plugins to a shared library.
