+++ src/alpaka/Makefile.deps
@@ -7,6 +7,7 @@ BeamSpotProducer_DEPENDS := Framework AlpakaCore AlpakaDataFormats DataFormats
 PixelTrackFitting_DEPENDS := Framework AlpakaCore AlpakaDataFormats
 PixelTriplets_DEPENDS := Framework AlpakaCore AlpakaDataFormats
 PixelVertexFinding_DEPENDS := Framework AlpakaCore AlpakaDataFormats DataFormats CondFormats
-SiPixelClusterizer_DEPENDS := Framework AlpakaCore AlpakaDataFormats DataFormats CondFormats
+SiPixelClusterizer_DEPENDS := Framework AlpakaCore AlpakaDataFormats DataFormats CondFormats SiPixelRawToDigi
+SiPixelRawToDigi_DEPENDS := DataFormats
 SiPixelRecHits_DEPENDS := Framework AlpakaCore AlpakaDataFormats DataFormats CondFormats
 Validation_DEPENDS := Framework AlpakaCore AlpakaDataFormats DataFormats
