+++ src/alpaka/plugins.txt
@@ -1,38 +1,38 @@
 BeamSpotESProducer pluginBeamSpotProducer.so
-alpaka_cuda_async::BeamSpotToAlpaka pluginBeamSpotProducer.so
-alpaka_tbb_async::BeamSpotToAlpaka pluginBeamSpotProducer.so
-alpaka_serial_sync::BeamSpotToAlpaka pluginBeamSpotProducer.so
-alpaka_cuda_async::CAHitNtupletAlpaka pluginPixelTriplets.so
-alpaka_tbb_async::CAHitNtupletAlpaka pluginPixelTriplets.so
-alpaka_serial_sync::CAHitNtupletAlpaka pluginPixelTriplets.so
-alpaka_cuda_async::PixelTrackSoAFromAlpaka pluginPixelTrackFitting.so
-alpaka_tbb_async::PixelTrackSoAFromAlpaka pluginPixelTrackFitting.so
-alpaka_serial_sync::PixelTrackSoAFromAlpaka pluginPixelTrackFitting.so
-alpaka_cuda_async::PixelVertexProducerAlpaka pluginPixelVertexFinding.so
-alpaka_tbb_async::PixelVertexProducerAlpaka pluginPixelVertexFinding.so
-alpaka_serial_sync::PixelVertexProducerAlpaka pluginPixelVertexFinding.so
-alpaka_cuda_async::PixelVertexSoAFromAlpaka pluginPixelVertexFinding.so
-alpaka_tbb_async::PixelVertexSoAFromAlpaka pluginPixelVertexFinding.so
-alpaka_serial_sync::PixelVertexSoAFromAlpaka pluginPixelVertexFinding.so
-alpaka_cuda_async::SiPixelRawToCluster pluginSiPixelClusterizer.so
-alpaka_tbb_async::SiPixelRawToCluster pluginSiPixelClusterizer.so
-alpaka_serial_sync::SiPixelRawToCluster pluginSiPixelClusterizer.so
+alpaka_cuda_async::BeamSpotToAlpaka pluginBeamSpotProducer_cuda.so
+alpaka_tbb_async::BeamSpotToAlpaka pluginBeamSpotProducer_tbb.so
+alpaka_serial_sync::BeamSpotToAlpaka pluginBeamSpotProducer_serial.so
+alpaka_cuda_async::CAHitNtupletAlpaka pluginPixelTriplets_cuda.so
+alpaka_tbb_async::CAHitNtupletAlpaka pluginPixelTriplets_tbb.so
+alpaka_serial_sync::CAHitNtupletAlpaka pluginPixelTriplets_serial.so
+alpaka_cuda_async::PixelTrackSoAFromAlpaka pluginPixelTrackFitting_cuda.so
+alpaka_tbb_async::PixelTrackSoAFromAlpaka pluginPixelTrackFitting_tbb.so
+alpaka_serial_sync::PixelTrackSoAFromAlpaka pluginPixelTrackFitting_serial.so
+alpaka_cuda_async::PixelVertexProducerAlpaka pluginPixelVertexFinding_cuda.so
+alpaka_tbb_async::PixelVertexProducerAlpaka pluginPixelVertexFinding_tbb.so
+alpaka_serial_sync::PixelVertexProducerAlpaka pluginPixelVertexFinding_serial.so
+alpaka_cuda_async::PixelVertexSoAFromAlpaka pluginPixelVertexFinding_cuda.so
+alpaka_tbb_async::PixelVertexSoAFromAlpaka pluginPixelVertexFinding_tbb.so
+alpaka_serial_sync::PixelVertexSoAFromAlpaka pluginPixelVertexFinding_serial.so
+alpaka_cuda_async::SiPixelRawToCluster pluginSiPixelClusterizer_cuda.so
+alpaka_tbb_async::SiPixelRawToCluster pluginSiPixelClusterizer_tbb.so
+alpaka_serial_sync::SiPixelRawToCluster pluginSiPixelClusterizer_serial.so
 SiPixelFedIdsESProducer pluginSiPixelClusterizer.so
-alpaka_cuda_async::SiPixelFedCablingMapESProducer pluginSiPixelClusterizer.so
-alpaka_tbb_async::SiPixelFedCablingMapESProducer pluginSiPixelClusterizer.so
-alpaka_serial_sync::SiPixelFedCablingMapESProducer pluginSiPixelClusterizer.so
-alpaka_cuda_async::SiPixelGainCalibrationForHLTESProducer pluginSiPixelClusterizer.so
-alpaka_tbb_async::SiPixelGainCalibrationForHLTESProducer pluginSiPixelClusterizer.so
-alpaka_serial_sync::SiPixelGainCalibrationForHLTESProducer pluginSiPixelClusterizer.so
-alpaka_cuda_async::PixelCPEFastESProducer pluginSiPixelRecHits.so
-alpaka_tbb_async::PixelCPEFastESProducer pluginSiPixelRecHits.so
-alpaka_serial_sync::PixelCPEFastESProducer pluginSiPixelRecHits.so
-alpaka_cuda_async::SiPixelRecHitAlpaka pluginSiPixelRecHits.so
-alpaka_tbb_async::SiPixelRecHitAlpaka pluginSiPixelRecHits.so
-alpaka_serial_sync::SiPixelRecHitAlpaka pluginSiPixelRecHits.so
-alpaka_cuda_async::CountValidator pluginValidation.so
-alpaka_tbb_async::CountValidator pluginValidation.so
-alpaka_serial_sync::CountValidator pluginValidation.so
-alpaka_cuda_async::HistoValidator pluginValidation.so
-alpaka_tbb_async::HistoValidator pluginValidation.so
-alpaka_serial_sync::HistoValidator pluginValidation.so
+alpaka_cuda_async::SiPixelFedCablingMapESProducer pluginSiPixelClusterizer_cuda.so
+alpaka_tbb_async::SiPixelFedCablingMapESProducer pluginSiPixelClusterizer_tbb.so
+alpaka_serial_sync::SiPixelFedCablingMapESProducer pluginSiPixelClusterizer_serial.so
+alpaka_cuda_async::SiPixelGainCalibrationForHLTESProducer pluginSiPixelClusterizer_cuda.so
+alpaka_tbb_async::SiPixelGainCalibrationForHLTESProducer pluginSiPixelClusterizer_tbb.so
+alpaka_serial_sync::SiPixelGainCalibrationForHLTESProducer pluginSiPixelClusterizer_serial.so
+alpaka_cuda_async::PixelCPEFastESProducer pluginSiPixelRecHits_cuda.so
+alpaka_tbb_async::PixelCPEFastESProducer pluginSiPixelRecHits_tbb.so
+alpaka_serial_sync::PixelCPEFastESProducer pluginSiPixelRecHits_serial.so
+alpaka_cuda_async::SiPixelRecHitAlpaka pluginSiPixelRecHits_cuda.so
+alpaka_tbb_async::SiPixelRecHitAlpaka pluginSiPixelRecHits_tbb.so
+alpaka_serial_sync::SiPixelRecHitAlpaka pluginSiPixelRecHits_serial.so
+alpaka_cuda_async::CountValidator pluginValidation_cuda.so
+alpaka_tbb_async::CountValidator pluginValidation_tbb.so
+alpaka_serial_sync::CountValidator pluginValidation_serial.so
+alpaka_cuda_async::HistoValidator pluginValidation_cuda.so
+alpaka_tbb_async::HistoValidator pluginValidation_tbb.so
+alpaka_serial_sync::HistoValidator pluginValidation_serial.so
